requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
filelock==3.15.4
Flask==3.0.3
fsspec==2024.6.1
h5py==3.11.0
huggingface-hub==0.24.5
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
mysql-connector==2.2.9
mysql-connector-python==9.0.0
networkx==3.3
numpy==1.26.4
packaging==24.1
pillow==10.4.0
pip==24.2
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.12.0
sentence-transformers==3.0.1
setuptools==72.1.0
six==1.16.0
smart-open==7.0.4
sympy==1.13.1
tbb==2021.13.1
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1+cpu
torchvision==0.18.1+cpu
tqdm==4.66.5
transformers==4.44.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
